[fix] modify boot script.

+++ Release.sh
@@ -59,7 +59,7 @@ elif [ "$1" == "factory" ]; then
 	echo "tar the file..." 
 	tar -czf Release.tar.gz *
 	echo "release is for factory preburn!" 
-else
+elif [ "$1" == "server" ]; then
 	rm ${workspace}/Release/STBCfg.xml
 	rm ${workspace}/Release/app.sh
 
@@ -67,7 +67,22 @@ else
 	SW_VER=${SW_VER_LINE#*\"}
 	SW_VER=${SW_VER%\"*}
 	echo "version is "${SW_VER}
-
-	echo ${SW_VER} > ${workspace}/Release/Program/version.dat
+	
+	echo "release is for usb disk copy!" 
+	cd ${workspace}/Release/Program/Lib/nexus/
+	sh ${workspace}/link2copy.sh
+	cd ${workspace}/Release/Program/Lib/common/
+	sh ${workspace}/link2copy.sh
+	cd ${workspace}/Release
+	echo "tar the Program file..."
+	tar -czf Program.tar.gz Program
+	
+	echo ${SW_VER} > client_version.txt
+	echo "check md5 sum..."
+	md5sum Program.tar.gz|cut -d ' ' -f1 >>  client_version.txt
+	
+	echo "remove Program/ and Private/"
+	rm -rf Program Private
+	
 	echo "release is for normal network mount update!" 
 fi
